WIP cleanup changes with server and client
