+++ asr/online_processor.py
@@ -1,0 +1,222 @@
+import sys
+import numpy as np
+
+import logging
+logger = logging.getLogger(__name__)
+
+from hypothesis_buffer import HypothesisBuffer
+
+class OnlineASRProcessor:
+
+    SAMPLING_RATE = 16000
+
+    def __init__(self, asr, tokenizer=None, buffer_trimming=("segment", 15), logfile=sys.stderr):
+        """asr: WhisperASR object
+        tokenizer: sentence tokenizer object for the target language. Must have a method *split* that behaves like the one of MosesTokenizer. It can be None, if "segment" buffer trimming option is used, then tokenizer is not used at all.
+        ("segment", 15)
+        buffer_trimming: a pair of (option, seconds), where option is either "sentence" or "segment", and seconds is a number. Buffer is trimmed if it is longer than "seconds" threshold. Default is the most recommended option.
+        logfile: where to store the log. 
+        """
+        self.asr = asr
+        self.tokenizer = tokenizer
+        self.logfile = logfile
+
+        self.init()
+
+        self.buffer_trimming_way, self.buffer_trimming_sec = buffer_trimming
+
+    def init(self, offset=None):
+        """run this when starting or restarting processing"""
+        self.audio_buffer = np.array([],dtype=np.float32)
+        self.transcript_buffer = HypothesisBuffer(logfile=self.logfile)
+        self.buffer_time_offset = 0
+        if offset is not None:
+            self.buffer_time_offset = offset
+        self.transcript_buffer.last_commited_time = self.buffer_time_offset
+        self.commited = []
+
+    def insert_audio_chunk(self, audio):
+        self.audio_buffer = np.append(self.audio_buffer, audio)
+
+    def prompt(self):
+        """Returns a tuple: (prompt, context), where "prompt" is a 200-character suffix of commited text that is inside of the scrolled away part of audio buffer. 
+        "context" is the commited text that is inside the audio buffer. It is transcribed again and skipped. It is returned only for debugging and logging reasons.
+        """
+        k = max(0,len(self.commited)-1)
+        while k > 0 and self.commited[k-1][1] > self.buffer_time_offset:
+            k -= 1
+
+        p = self.commited[:k]
+        p = [t for _,_,t,_ in p]  # Extract text from 4-tuples
+        prompt = []
+        l = 0
+        while p and l < 200:  # 200 characters prompt size
+            x = p.pop(-1)
+            l += len(x)+1
+            prompt.append(x)
+        non_prompt = self.commited[k:]
+        return self.asr.sep.join(prompt[::-1]), self.asr.sep.join(t for _,_,t,_ in non_prompt)
+
+    def process_iter(self):
+        """Runs on the current audio buffer.
+        Returns: a tuple (beg_timestamp, end_timestamp, "text"), or (None, None, ""). 
+        The non-emty text is confirmed (committed) partial transcript.
+        """
+
+        prompt, non_prompt = self.prompt()
+        logger.debug(f"PROMPT: {prompt}")
+        logger.debug(f"CONTEXT: {non_prompt}")
+        logger.debug(f"transcribing {len(self.audio_buffer)/self.SAMPLING_RATE:2.2f} seconds from {self.buffer_time_offset:2.2f}")
+        res = self.asr.transcribe(self.audio_buffer, init_prompt=prompt)
+
+        # transform to [(beg,end,"word1"), ...]
+        tsw = self.asr.ts_words(res)
+
+        self.transcript_buffer.insert(tsw, self.buffer_time_offset)
+        o = self.transcript_buffer.flush()
+        self.commited.extend(o)
+        completed = self.to_flush(o)
+        logger.debug(f">>>>COMPLETE NOW: {completed}")
+        the_rest = self.to_flush(self.transcript_buffer.complete())
+        logger.debug(f"INCOMPLETE: {the_rest}")
+
+        # there is a newly confirmed text
+
+        if o and self.buffer_trimming_way == "sentence":  # trim the completed sentences
+            if len(self.audio_buffer)/self.SAMPLING_RATE > self.buffer_trimming_sec:  # longer than this
+                self.chunk_completed_sentence()
+
+        
+        if self.buffer_trimming_way == "segment":
+            s = self.buffer_trimming_sec  # trim the completed segments longer than s,
+        else:
+            s = 30 # if the audio buffer is longer than 30s, trim it
+        
+        if len(self.audio_buffer)/self.SAMPLING_RATE > s:
+            self.chunk_completed_segment(res)
+
+            # alternative: on any word
+            #l = self.buffer_time_offset + len(self.audio_buffer)/self.SAMPLING_RATE - 10
+            # let's find commited word that is less
+            #k = len(self.commited)-1
+            #while k>0 and self.commited[k][1] > l:
+            #    k -= 1
+            #t = self.commited[k][1] 
+            logger.debug("chunking segment")
+            #self.chunk_at(t)
+
+        logger.debug(f"len of buffer now: {len(self.audio_buffer)/self.SAMPLING_RATE:2.2f}")
+        return self.to_flush(o)
+
+    def chunk_completed_sentence(self):
+        if self.commited == []: return
+        logger.debug(self.commited)
+        sents = self.words_to_sentences(self.commited)
+        for s in sents:
+            logger.debug(f"\t\tSENT: {s}")
+        if len(sents) < 2:
+            return
+        while len(sents) > 2:
+            sents.pop(0)
+        # we will continue with audio processing at this timestamp
+        chunk_at = sents[-2][1]
+
+        logger.debug(f"--- sentence chunked at {chunk_at:2.2f}")
+        self.chunk_at(chunk_at)
+
+    def chunk_completed_segment(self, res):
+        if self.commited == []: return
+
+        ends = self.asr.segments_end_ts(res)
+
+        t = self.commited[-1][1]
+
+        if len(ends) > 1:
+
+            e = ends[-2]+self.buffer_time_offset
+            while len(ends) > 2 and e > t:
+                ends.pop(-1)
+                e = ends[-2]+self.buffer_time_offset
+            if e <= t:
+                logger.debug(f"--- segment chunked at {e:2.2f}")
+                self.chunk_at(e)
+            else:
+                logger.debug(f"--- last segment not within commited area")
+        else:
+            logger.debug(f"--- not enough segments to chunk")
+
+
+
+
+
+    def chunk_at(self, time):
+        """trims the hypothesis and audio buffer at "time"
+        """
+        self.transcript_buffer.pop_commited(time)
+        cut_seconds = time - self.buffer_time_offset
+        self.audio_buffer = self.audio_buffer[int(cut_seconds*self.SAMPLING_RATE):]
+        self.buffer_time_offset = time
+
+    def words_to_sentences(self, words):
+        """Uses self.tokenizer for sentence segmentation of words.
+        Returns: [(beg,end,"sentence 1",avg_prob),...]
+        """
+        
+        cwords = [w for w in words]
+        t = " ".join(o[2] for o in cwords)
+        s = self.tokenizer.split(t)
+        out = []
+        while s:
+            beg = None
+            end = None
+            sent = s.pop(0).strip()
+            fsent = sent
+            sent_probs = []
+            while cwords:
+                b,e,w,p = cwords.pop(0)
+                w = w.strip()
+                if beg is None and sent.startswith(w):
+                    beg = b
+                    sent_probs.append(p)
+                elif end is None and sent == w:
+                    end = e
+                    sent_probs.append(p)
+                    avg_prob = sum(sent_probs) / len(sent_probs) if sent_probs else 1.0
+                    out.append((beg,end,fsent,avg_prob))
+                    break
+                else:
+                    sent_probs.append(p)
+                sent = sent[len(w):].strip()
+        return out
+
+    def finish(self):
+        """Flush the incomplete text when the whole processing ends.
+        Returns: the same format as self.process_iter()
+        """
+        o = self.transcript_buffer.complete()
+        f = self.to_flush(o)
+        logger.debug(f"last, noncommited: {f}")
+        self.buffer_time_offset += len(self.audio_buffer)/16000
+        return f
+
+
+    def to_flush(self, sents, sep=None, offset=0, ):
+        # concatenates the timestamped words or sentences into one sequence that is flushed in one line
+        # sents: [(beg1, end1, "sentence1", prob1), ...] or [] if empty
+        # return: (beg1,end-of-last-sentence,"concatenation of sentences", avg_prob, word_probabilities) or (None, None, "", 1.0, []) if empty
+        if sep is None:
+            sep = self.asr.sep
+        t = sep.join(s[2] for s in sents)
+        if len(sents) == 0:
+            b = None
+            e = None
+            avg_prob = 1.0
+            word_probs = []
+        else:
+            b = offset + sents[0][0]
+            e = offset + sents[-1][1]
+            # Calculate average probability
+            avg_prob = sum(s[3] for s in sents) / len(sents) if sents else 1.0
+            # Create word probability list: [(word, probability), ...]
+            word_probs = [(s[2].strip(), s[3]) for s in sents if s[2].strip()]
+        return (b,e,t,avg_prob,word_probs)
